+++ test/service/ConfigLoader.test.ts
@@ -17,9 +17,10 @@
 import { expect } from '@oclif/test';
 import 'mocha';
 import { Account, NetworkType } from 'symbol-sdk';
+import { LoggerFactory, LogType } from '../../src';
 import { ConfigAccount, PrivateKeySecurityMode } from '../../src/model';
 import { BootstrapUtils, ConfigLoader, KeyName, Preset } from '../../src/service';
-
+const logger = LoggerFactory.getLogger(LogType.Silent);
 class ConfigLoaderMocked extends ConfigLoader {
     public generateAccount = (
         networkType: NetworkType,
@@ -33,7 +34,7 @@ class ConfigLoaderMocked extends ConfigLoader {
 
 describe('ConfigLoader', () => {
     it('ConfigLoader loadPresetData testnet no assembly', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         try {
             await configLoader.createPresetData({
                 preset: Preset.testnet,
@@ -50,7 +51,7 @@ describe('ConfigLoader', () => {
     });
 
     it('ConfigLoader loadPresetData testnet assembly', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const presetData = await configLoader.createPresetData({
             preset: Preset.testnet,
             assembly: 'dual',
@@ -62,7 +63,7 @@ describe('ConfigLoader', () => {
     });
 
     it('ConfigLoader loadPresetData bootstrap custom', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const presetData = await configLoader.createPresetData({
             preset: Preset.bootstrap,
             assembly: undefined,
@@ -77,7 +78,7 @@ describe('ConfigLoader', () => {
     });
 
     it('ConfigLoader loadPresetData bootstrap custom too short!', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         try {
             await configLoader.createPresetData({
                 preset: Preset.bootstrap,
@@ -92,7 +93,7 @@ describe('ConfigLoader', () => {
     });
 
     it('applyIndex', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const context = { $index: 10 };
         expect(configLoader.applyValueTemplate(context, 'hello')).to.be.eq('hello');
         expect(configLoader.applyValueTemplate(context, 'index')).to.be.eq('index');
@@ -105,7 +106,7 @@ describe('ConfigLoader', () => {
     });
 
     it('expandServicesRepeat when repeat 3', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const services = [
             {
                 repeat: 3,
@@ -158,7 +159,7 @@ describe('ConfigLoader', () => {
     });
 
     it('expandServicesRepeat when repeat 0', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const services = [
             {
                 repeat: 0,
@@ -179,7 +180,7 @@ describe('ConfigLoader', () => {
     });
 
     it('expandServicesRepeat when no repeat', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const services = [
             {
                 apiNodeName: 'api-node-{{$index}}',
@@ -211,7 +212,7 @@ describe('ConfigLoader', () => {
     });
 
     it('applyValueTemplate when object', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const value = {
             _info: 'this file contains a list of api-node peers',
             knownPeers: [
@@ -234,7 +235,7 @@ describe('ConfigLoader', () => {
     });
 
     it('applyValueTemplate when array', async () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const value = [
             {
                 _info: 'this file contains a list of api-node peers',
@@ -259,7 +260,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should migrated old addresses', () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const oldAddresses = BootstrapUtils.loadYaml('./test/addresses/addresses-old.yml', false);
         const newAddresses = BootstrapUtils.loadYaml('./test/addresses/addresses-new.yml', false);
         const addresses = configLoader.migrateAddresses(oldAddresses, NetworkType.TEST_NET);
@@ -267,14 +268,14 @@ describe('ConfigLoader', () => {
     });
 
     it('should migrated not migrate new addresses', () => {
-        const configLoader = new ConfigLoaderMocked();
+        const configLoader = new ConfigLoaderMocked(logger);
         const newAddresses = BootstrapUtils.loadYaml('./test/addresses/addresses-new.yml', false);
         const addresses = configLoader.migrateAddresses(newAddresses, NetworkType.TEST_NET);
         expect(addresses).to.be.deep.eq(newAddresses);
     });
 
     it('should generateAccount when old and new are different', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -299,7 +300,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are different', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -324,7 +325,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are same', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -349,7 +350,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are same, new no private eky', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -374,7 +375,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are same, old no private eky', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -398,7 +399,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are same, old private key', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -421,7 +422,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are different, no private key new', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -445,7 +446,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are different. No new account', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -469,7 +470,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount when old and new are different. No new account. Old without private', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const oldAccount = Account.generateNewAccount(networkType);
@@ -491,7 +492,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount brand new', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.ENCRYPT;
         const account = configLoader.generateAccount(networkType, securityMode, KeyName.Main, undefined, undefined, undefined);
@@ -501,7 +502,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount brand new on remote', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.PROMPT_MAIN_TRANSPORT;
         const account = configLoader.generateAccount(networkType, securityMode, KeyName.Remote, undefined, undefined, undefined);
@@ -511,7 +512,7 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount brand new on voting', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.PROMPT_MAIN;
         const account = configLoader.generateAccount(networkType, securityMode, KeyName.Voting, undefined, undefined, undefined);
@@ -521,21 +522,21 @@ describe('ConfigLoader', () => {
     });
 
     it('should generateAccount raise error new', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.PROMPT_MAIN;
         expect(() => configLoader.generateAccount(networkType, securityMode, KeyName.Main, undefined, undefined, undefined)).throw;
     });
 
     it('should generateAccount raise error new', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.PROMPT_MAIN_TRANSPORT;
         expect(() => configLoader.generateAccount(networkType, securityMode, KeyName.Transport, undefined, undefined, undefined)).throw;
     });
 
     it('should generateAccount raise error new', () => {
-        const configLoader = new ConfigLoader();
+        const configLoader = new ConfigLoader(logger);
         const networkType = NetworkType.TEST_NET;
         const securityMode = PrivateKeySecurityMode.PROMPT_ALL;
         expect(() => configLoader.generateAccount(networkType, securityMode, KeyName.Remote, undefined, undefined, undefined)).throw;