+++ test/service/ConfigService.test.ts
@@ -16,11 +16,12 @@
 
 import { expect } from '@oclif/test';
 import 'mocha';
+import { LoggerFactory, LogType } from '../../src';
 import { ConfigService, CryptoUtils, Preset } from '../../src/service';
-
+const logger = LoggerFactory.getLogger(LogType.Silent);
 describe('ConfigService', () => {
     it('ConfigService default run with optin_preset.yml', async () => {
-        await new ConfigService({
+        await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/ConfigService.test.optin',
@@ -29,7 +30,7 @@ describe('ConfigService', () => {
     });
 
     it('ConfigService default run with override-currency-preset.yml', async () => {
-        await new ConfigService({
+        await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/ConfigService.test.custom',
@@ -38,7 +39,7 @@ describe('ConfigService', () => {
     });
 
     it('ConfigService testnet assembly', async () => {
-        await new ConfigService({
+        await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/ConfigService.test.testnet',
@@ -48,7 +49,7 @@ describe('ConfigService', () => {
     });
 
     it('ConfigService mainnet assembly', async () => {
-        await new ConfigService({
+        await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/ConfigService.test.mainnet',
@@ -58,14 +59,14 @@ describe('ConfigService', () => {
     });
 
     it('ConfigService bootstrap default', async () => {
-        const configResult = await new ConfigService({
+        const configResult = await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/bootstrap',
             preset: Preset.bootstrap,
         }).run();
 
-        const configResultUpgrade = await new ConfigService({
+        const configResultUpgrade = await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             upgrade: true,
             target: 'target/tests/bootstrap',
@@ -78,7 +79,7 @@ describe('ConfigService', () => {
     });
 
     it('ConfigService bootstrap repeat', async () => {
-        const configResult = await new ConfigService({
+        const configResult = await new ConfigService(logger, {
             ...ConfigService.defaultParams,
             reset: true,
             target: 'target/tests/ConfigService.bootstrap.repeat',